requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1
boto3==1.17.49
botocore==1.20.49; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
dj-database-url==0.5.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.11.1
django-tinymce==3.2.0
django==3.1.7
gunicorn==20.0.4
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pillow==8.1.2
psycopg2-binary==2.8.6
psycopg2==2.8.6
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
s3transfer==0.3.6
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
whitenoise==5.2.0
